requirements.txt
flask
python-dotenv
pandas
requests
pillow
tensorflow-cpu